requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
langchain==0.1.0
langchain-google-genai==0.0.6
chromadb==0.4.22
sentence-transformers==2.2.2
pydantic==2.5.0
numpy==1.26.4
requests==2.31.0
setuptools==69.0.0
